Extract shared Windows sandbox session runner
