+++ codex-rs/windows-sandbox-rs/Cargo.toml
@@ -37,7 +37,7 @@ glob = { workspace = true }
 serde = { version = "1.0", features = ["derive"] }
 serde_json = "1.0"
 tempfile = "3"
-tokio = { workspace = true, features = ["sync", "rt"] }
+tokio = { workspace = true, features = ["sync", "rt", "signal", "time"] }
 tracing-appender = { workspace = true }
 windows = { version = "0.58", features = [
     "Win32_Foundation",
